Merge C++ changes from 1.4

+++ DeviceAdapters/IIDC/IIDCConvert.h
@@ -2,7 +2,8 @@
 //
 // AUTHOR:        Mark A. Tsuchida
 //
-// COPYRIGHT:     University of California, San Francisco, 2015
+// COPYRIGHT:     2014-2015, Regents of the University of California
+//                2016, Open Imaging, Inc.
 //
 // LICENSE:       This file is distributed under the BSD license.
 //                License text is included with the source distribution.
@@ -44,7 +45,8 @@ ConvertToRGB8(uint8_t* dst, const uint8_t* src, size_t width, size_t height,
       default:
          return;
    }
-   dc1394_convert_to_RGB8(const_cast<uint8_t*>(src), dst, width, height,
+   dc1394_convert_to_RGB8(const_cast<uint8_t*>(src), dst,
+         static_cast<uint32_t>(width), static_cast<uint32_t>(height),
          DC1394_BYTE_ORDER_UYVY, coding, 8);
 }
